Fix new operation not being started after previous is cancelled

+++ UnrealAutomationCommon/Operations/OperationRunner.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading;
+﻿using System;
+using System.Threading;
 using System.Threading.Tasks;
 using UnrealAutomationCommon.Operations.BaseOperations;
 using UnrealAutomationCommon.Operations.OperationOptionTypes;
@@ -13,11 +14,16 @@ namespace UnrealAutomationCommon.Operations
         private readonly OperationParameters _operationParameters;
         private int _lineCount;
 
+        //public event Action Ended;
+        private Task<OperationResult> _currentTask = null;
+
         public OperationRunner(Operation operation, OperationParameters operationParameters)
         {
             Operation = operation;
             _operationParameters = operationParameters;
         }
+
+        public bool IsRunning => _currentTask is { IsCompleted: false };
 
         public Operation Operation { get; }
 
@@ -30,21 +36,42 @@ namespace UnrealAutomationCommon.Operations
 
         public async Task<OperationResult> Run()
         {
+            if (_currentTask != null)
+            {
+                throw new Exception("Task is already running");
+            }
+
             string outputPath = Operation.GetOutputPath(_operationParameters);
             FileUtils.DeleteDirectoryIfExists(outputPath);
 
-            var task = Operation.Execute(_operationParameters, this, _cancellationTokenSource.Token);
+            _currentTask = Operation.Execute(_operationParameters, this, _cancellationTokenSource.Token);
 
             FlagOptions flagOptions = _operationParameters.FindOptions<FlagOptions>();
             if (flagOptions is { WaitForAttach: true }) Output?.Invoke("-WaitForAttach was specified, attach now", LogVerbosity.Log);
 
-            return await task;
+            OperationResult result = await _currentTask;
+            Output?.Invoke($"'{Operation.OperationName}' task ended", LogVerbosity.Log);
+            _currentTask = null;
+            return result;
         }
 
-        public void Cancel()
+        public async Task Cancel()
         {
-            Output?.Invoke("Cancelling operation '" + Operation.OperationName + "'", LogVerbosity.Warning);
+            if (_currentTask == null)
+            {
+                throw new Exception("Task is not running");
+            }
+
+            Output?.Invoke($"Cancelling operation '{Operation.OperationName}'", LogVerbosity.Warning);
+
+            //TaskCompletionSource<bool> tcs = new();
+            //Ended += () => tcs.TrySetResult(true);
+
             _cancellationTokenSource.Cancel();
+
+            await _currentTask;
+
+            Output?.Invoke($"'{Operation.OperationName}' task ended from cancellation", LogVerbosity.Warning);
         }
 
         private void OutputLine(string line, LogVerbosity verbosity)
